Add files via upload
A program I made a few years ago when I first dabbled in learning very basic C++, before I began taking programming too seriously. It's admittedly far from my best work, but still something I wanted to share on here regardless. This is an IGT (in-game time) calculator for Super Smash Bros. Melee Target Test speedrunners. I made it to practice simple I/O programs in the command line, and to get used to the general syntax of C++. I can tell just from looking at it now that there is quite a bit I'd change if I were to rewrite it, and at some point I may, but for now I will simply display it as is. For best functionality, compile on Windows (I know, the Windows system commands are terrible but at the time I was willing to settle). Will work as is on Linux, and I would assume Mac as well, it's just a little uglier.